Refactor in order to support return url
In order to support returnUrl as specified in BankID docs we now expose our transaction_id to be used as nonce in the returnUrl.
See https://developers.bankid.com/getting-started/frontend/return-url

- Remove signing/unsigning of transaction_id
I seen no reason for validating the 5 minute lifetime of auth orders. BankID already handles this on their side. Our transaction_id also isn't secret for the user logging in, it's our reference to the BankID auth order.